refactor: Move TableFormatter to framework.text package
