+++ src/main/java/org/opentripplanner/framework/text/TableFormatter.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.util.lang;
+package org.opentripplanner.framework.text;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -8,6 +8,7 @@ import java.util.Objects;
 import java.util.function.BiFunction;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.opentripplanner.util.lang.StringUtils;
 
 /**
  * This class is responsible for creating a pretty table that can be printed to a terminal window.
