+++ src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/path/PathDiff.java
@@ -1,8 +1,8 @@
 package org.opentripplanner.routing.algorithm.raptoradapter.path;
 
-import static org.opentripplanner.util.lang.TableFormatter.Align.Center;
-import static org.opentripplanner.util.lang.TableFormatter.Align.Left;
-import static org.opentripplanner.util.lang.TableFormatter.Align.Right;
+import static org.opentripplanner.framework.text.TableFormatter.Align.Center;
+import static org.opentripplanner.framework.text.TableFormatter.Align.Left;
+import static org.opentripplanner.framework.text.TableFormatter.Align.Right;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -10,13 +10,13 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
+import org.opentripplanner.framework.text.TableFormatter;
 import org.opentripplanner.routing.util.DiffEntry;
 import org.opentripplanner.routing.util.DiffTool;
 import org.opentripplanner.transit.raptor.api.path.Path;
 import org.opentripplanner.transit.raptor.api.path.PathLeg;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
 import org.opentripplanner.util.CompositeComparator;
-import org.opentripplanner.util.lang.TableFormatter;
 import org.opentripplanner.util.time.DurationUtils;
 import org.opentripplanner.util.time.TimeUtils;
