+++ src/test/java/org/opentripplanner/generate/doc/framework/ConfigTypeTable.java
@@ -6,8 +6,8 @@ import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.framework.text.MarkdownFormatter;
+import org.opentripplanner.framework.text.TableFormatter;
 import org.opentripplanner.standalone.config.framework.json.ConfigType;
-import org.opentripplanner.util.lang.TableFormatter;
 
 @SuppressWarnings("NewClassNamingConvention")
 public class ConfigTypeTable {
