+++ src/test/java/org/opentripplanner/transit/raptor/RaptorArchitectureTest.java
@@ -1,6 +1,7 @@
 package org.opentripplanner.transit.raptor;
 
 import static com.tngtech.archunit.library.dependencies.SlicesRuleDefinition.slices;
+import static org.opentripplanner.OtpArchitectureModules.FRAMEWORK_TEXT;
 import static org.opentripplanner.OtpArchitectureModules.GNU_TROVE;
 import static org.opentripplanner.OtpArchitectureModules.RAPTOR_API;
 import static org.opentripplanner.OtpArchitectureModules.TRANSIT;
@@ -53,7 +54,7 @@ public class RaptorArchitectureTest {
     var internalApi = RR_INTERNAL_API.dependsOn(RAPTOR_API).verify();
 
     // RangeRaptor common allowed dependencies
-    var common = Module.of(UTILS, GNU_TROVE, RAPTOR_API, RAPTOR_UTILS, internalApi);
+    var common = Module.of(FRAMEWORK_TEXT, UTILS, GNU_TROVE, RAPTOR_API, RAPTOR_UTILS, internalApi);
 
     var debug = rr.subPackage("debug").dependsOn(common).verify();
     var lifecycle = rr.subPackage("lifecycle").dependsOn(common).verify();
