+++ src/test/java/org/opentripplanner/util/lang/TableFormatterTest.java
@@ -1,13 +1,14 @@
 package org.opentripplanner.util.lang;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.opentripplanner.util.lang.TableFormatter.Align.Center;
-import static org.opentripplanner.util.lang.TableFormatter.Align.Left;
-import static org.opentripplanner.util.lang.TableFormatter.Align.Right;
+import static org.opentripplanner.framework.text.TableFormatter.Align.Center;
+import static org.opentripplanner.framework.text.TableFormatter.Align.Left;
+import static org.opentripplanner.framework.text.TableFormatter.Align.Right;
 
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.framework.text.TableFormatter;
 
 public class TableFormatterTest {
